requirements.txt
fastapi>=0.115.0
uvicorn>=0.23.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.10.0 
boto3==1.35.0
motor>=3.4.0
celery==5.2.7
redis==4.5.4
flower==1.2.0
python-multipart==0.0.20